% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_FisherPValue_OnePair.R
\name{DESeq2_FisherPvalue_Onepair}
\alias{DESeq2_FisherPvalue_Onepair}
\title{Combine P values for one pair of input and output}
\usage{
DESeq2_FisherPvalue_Onepair(filepath, rawdatafile, outputname.feature,
  outputname.oligo, num.in, num.out, name.out)
}
\arguments{
\item{filepath}{The path of the raw data (excluding the data file itself). Remember to double quote the filepath.}

\item{rawdatafile}{The name of the raw data file. Remember to double quote the rawdatafile, and add the ".xlsx" after the name of rawdatafile (It must be excel file)!}

\item{outputname.feature}{The name of output file for each gene you want.}

\item{outputname.oligo}{The name of output file for each oligo you want.}

\item{num.in}{The number of input columns in the raw data file.}

\item{num.out}{The number of output columns in the raw data file.}

\item{name.out}{The prefix of columns in the output. For example, "A vs B" or "B vs C", in which "A". "B" and "C" represent the names of input and output in the raw data file.}
}
\description{
This function is used to: 
1. To the raw data, using DESeq2 to give the p values of input vesus output
2. Combine the p values in the output of DESeq2
3. To the raw data, pre-sum the count data group by locus_tag, and using DESeq2 to give the p values of input vesus output
}
\details{
This function should be used on a raw data with the following layout:
1. The first three columns must be "Feature", "Locus_Tag", "Strand"
   Here "Feture" is like "A000XXX". Be sure that the colnames are right!
2. The next columns are for input and output successively. 
   Note that there is no constraint for the names of these input and
   output columns, but there can only be 1 kind of output. The number
   of columns in each output can be any number.
}
\examples{
DESeq2_FisherPvalue_Onepair(filepath = "C:/Users/", rawdatafile = "rawdata.xlsx", outputname = "output", num.in = 3, num.out = 3, name.out = "A vs B")
}
\keyword{mmcc}
