% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_FisherPValue_TwoPair.R
\name{DESeq2_FisherPvalue_Twopair}
\alias{DESeq2_FisherPvalue_Twopair}
\title{Combine P values for raw data with one input and two outputs}
\usage{
DESeq2_FisherPvalue_Twopair(filepath, rawdatafile, outputname, num.in, num.out1,
  num.out2, name.out1, name.out2)
}
\arguments{
\item{filepath}{The path of the raw data (excluding the data file itself). Remember to double quote the filepath.}

\item{rawdatafile}{The name of the raw data file. Remember to double quote the rawdatafile, and add the ".xlsx" after the name of rawdatafile (It must be excel file)!}

\item{outputname}{The name of output file you want.}

\item{num.in}{The number of input columns in the raw data file.}

\item{num.out1}{The number of output1 columns in the raw data file.}

\item{num.out2}{The number of output2 columns in the raw data file.}

\item{name.out1}{The prefix of output1 columns in the output file. For example, if "A" is input, "B1" is output1, "B2" is output2, then here "B1" can be used.}

\item{name.out2}{The prefix of output2 columns in the output file. For example, if "A" is input, "B1" is output1, "B2" is output2, then here "B2" can be used.}
}
\description{
This function is a variation of function \link[mmcc]{DESeq2_FisherPvalue_Onepair}. Now two different types of output are allowed.
}
\details{
This function should be used on a raw data with the following layout:
1. The first three columns must be "Feature", "Locus_Tag", "Strand"
   Here "Feture" is like "A000XXX". Be sure that the colnames are right!
2. The next columns are for input, output1, output2 successively. 
   Note that there is no constraint for the names of these input and
   output columns, but there can only be 2 kinds of outputs. The number
   of columns in each output can be any number.
}
\examples{
DESeq2_FisherPvalue_Twopair(filepath = "C:/Users/", rawdatafile = "rawdata.xlsx", outputname = "output", num.in = 3, num.out1 = 3, num.out2 = 3, name.out1 = "B1", name.out2 = "B2")
}
\keyword{mmcc}
