% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_autocorr.R
\name{diag_autocorr}
\alias{diag_autocorr}
\title{Diagnostic: Calculate the autocorrelation for each chain and parameter}
\usage{
diag_autocorr(x, lags = NULL)
}
\arguments{
\item{x}{mcmc.list object or dataframe creatd by \code{mcmc_to_dt}}

\item{lags}{integer the lag value that you want to use}
}
\value{
a data.table data frame with the autocorrelation at each lag for each chain and parameter columns \code{lag}, \code{acf}, \code{chain}, and \code{parameter}
}
\description{
Diagnostic: Calculate the autocorrelation for each chain and parameter
}
\examples{
library(coda)
data(line)
line_acf <- diag_autocorr(line)
line_acf
}
