% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-batchmeans.R
\name{diag_mc_stderr}
\alias{diag_mc_stderr}
\title{Batch Means Standard Errors}
\usage{
diag_mc_stderr(x, b_size = "sqroot", warn = TRUE)
}
\arguments{
\item{x}{'mcmc.list' or 'data.table' object}

\item{b_size}{character, method for determining size of batch (see Details)}

\item{warn}{logical, give a warning of there are too few samples in the MCMC output}
}
\description{
Calculate Monte Carlo Standard Errors Using Batch Means
}
\details{
For the batch size, the options are 'sqroot', for using the using the square root of the MCMC output length, or 'cuberoot', for using the cube root of the output length. The default is 'sqroot'.
}
\examples{
library(coda)
data(line)
r <- diag_mc_stderr(line)
r

}
\references{
Galin L. Jones, Murali Haran, Brian S. Caffo, and Ronald Neath (2006). "Fixed-Width Output Analysis for Markov Chain Monte Carlo," Journal of the American Statistical Association, 101, 1537--1547
}
