% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-jags-model.R
\docType{data}
\name{example_jags_model}
\alias{example_jags_model}
\title{Example JAGS Model}
\format{An object of class \code{jags} of length 8.}
\usage{
data(example_jags_model)
}
\description{
This is an example JAGS model to use for examples in the `mmcc` package. The
    model is fit as a basic linear regression, with uniform priors on beta0
    and beta1, where y is assumed to be normal, with mean mu and precision
    tau. The model is drawn from the vignette, "Model summaries for a
    Bayesian linear regression", and the code to generate it can be found in
    the `data-raw` folder.
}
\examples{

glance(example_jags_model)
}
\keyword{data}
