% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-stan-model.R
\docType{data}
\name{example_stan_model}
\alias{example_stan_model}
\title{Example STAN Model}
\format{An object of class \code{stanfit} of dimension 5 x 4 x 3.}
\usage{
data(example_stan_model)
}
\description{
This is an example STAN model to use for examples in the `mmcc` package. It
    was created with the example code given below in examples.
}
\examples{
\dontrun{
    library(rstan)
    scode <- "
    parameters {
    real y[2];
    }
    model {
    y[1] ~ normal(0, 1);
    y[2] ~ double_exponential(0, 2);
    }
    "
    example_stan_model <- stan(model_code = scode,
                               iter = 10,
                               verbose = FALSE)
}

glance(example_stan_model)
}
\keyword{data}
