% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.dic.R
\name{glance.dic}
\alias{glance.dic}
\title{Glance upon your dic samples to get the penalised deviance}
\usage{
\method{glance}{dic}(x, ...)
}
\arguments{
\item{x}{a dic object generated by rjags::dic.samples}

\item{...}{(optional) additional arguments to pass}
}
\value{
a one-row dataframe of summary information of the dic samples
}
\description{
This provides a one-row dataframe (\code{data.table}) with information on the
overall deviance, effective number of parameters (when \code{type} was pD) or
optimism (when \code{type} was popt), and resulting penalised deviance.
}
\examples{
library(rjags)
example_jags_model$recompile()
model_dic <- dic.samples(example_jags_model, n.iter = 1000)
glance(model_dic)
}
\author{
Sam Clifford, \email{sj.clifford@gmail.com}
}
