% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance-mcmc-list.R
\name{glance.mcmc.list}
\alias{glance.mcmc.list}
\title{Glance upon your mcmclist to get summary information}
\usage{
\method{glance}{mcmc.list}(x, ...)
}
\arguments{
\item{x}{an mcmc.list object}

\item{...}{(optional) additional arguments to pass}
}
\value{
a one-row dataframe of summary information of the mcmc model
}
\description{
This provides a one-row dataframe with information on number of chains, the
    number of variables, the number of iterations, and the lower and upper
    values for effective sample size (ess), and rhat.
}
\examples{

library(coda)
data(line)
glance(line)

}
