% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-chain.R, R/n-var.R
\name{mcmc-dims}
\alias{mcmc-dims}
\alias{n_chain}
\alias{n_var}
\title{Dimensions of MCMC objects}
\usage{
n_chain(x)

n_var(x)
}
\arguments{
\item{x}{an mcmc object, see "Details" for a list of supported mcmc objects.}
}
\description{
Retrieve dimensions of an MCMC object:
\itemize{
\item \code{n_chain(x)} the number of chains
\item \code{n_var(x)} the number of variables
\item \code{n_iter(x)} the number of iterations
}
}
\details{
This is similar to \code{coda::nchain()} but more general, working for
the following classes:
* \code{mcmc}
* \code{mcmc.list}
* \code{data.table} (generated by \code{mcmc_to_dt})
* \code{stanfit}
* \code{jags}
}
