% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tidier_stan.R
\name{mcmc_to_dt.stanfit}
\alias{mcmc_to_dt.stanfit}
\title{MCMC tidiers (draft) for STAN}
\usage{
\method{mcmc_to_dt}{stanfit}(mcmc_object, ...)
}
\arguments{
\item{mcmc_object}{an object of class "mcmc.list", as you would find with
fitting a model using `jags.model()`, and `coda.samples`, or "stan", from
fitting a stan model.}

\item{...}{additional arguments}
}
\description{
MCMC tidiers (draft) for STAN
}
\examples{
 mcmc_to_dt(example_stan_model)

}
