% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tidier_stan.R
\name{mcmc_to_dt_stan}
\alias{mcmc_to_dt_stan}
\title{MCMC tidiers (draft) for STAN}
\usage{
mcmc_to_dt_stan(stan_model)
}
\arguments{
\item{stan_model}{a stan model}
}
\value{
a tidy dataframe
}
\description{
MCMC tidiers (draft) for STAN
}
\examples{
\dontrun{
#### example 1
library(rstan)
scode <- "
parameters {
real y[2];
}
model {
y[1] ~ normal(0, 1);
y[2] ~ double_exponential(0, 2);
}
"
fit1 <- stan(model_code = scode, iter = 10, verbose = FALSE)
 mcmc_to_dt_stan(fit1)
 }


}
