% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-helpers.R
\name{n_chain}
\alias{n_chain}
\title{Dimensions of MCMC objects}
\usage{
n_chain(x)
}
\description{
Retrieve dimensions of an MCMC object:

 * `n_chain(x)` the number of chains
 * `n_var(x)` the number of variables.
 * `n_iter(x)` the number of iterations
}
