% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_sims.R
\name{n_sims}
\alias{n_sims}
\title{Return the number of simulations}
\usage{
n_sims(x)
}
\arguments{
\item{x}{mcmc list}
}
\value{
integer of number of simulations
}
\description{
Return the number of simulations
}
\author{
Nicholas Tierney
}
