% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tidiers.R
\name{tidy_mcmc}
\alias{tidy_mcmc}
\title{tidy_mcmc}
\usage{
tidy_mcmc(mcmc_object)
}
\arguments{
\item{mcmc_object}{an object of class "mcmc.list", as you would find with fitting a model using `jags.model()`, and `coda.samples`}
}
\value{
a data.table dataframe
}
\description{
use data.table to return a tidy dataframe from an "mcmc.list" object
}

