% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example-stan-model.R
\docType{data}
\name{example_stan_model}
\alias{example_stan_model}
\title{Example STAN Model}
\format{
An object of class \code{stanfit} of dimension 5 x 4 x 3.
}
\usage{
data(example_stan_model)
}
\description{
This is an example STAN model to use for examples in the \code{mmcc} package. It
was created with the example code given below in examples.
}
\examples{

mcmc_to_dt(example_stan_model)
}
\keyword{data}
