% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.dic.R
\name{glance.dic}
\alias{glance.dic}
\title{Glance upon your dic samples to get the penalised deviance}
\usage{
\method{glance}{dic}(x)
}
\arguments{
\item{x}{a dic object generated by rjags::dic.samples}
}
\value{
a one-row dataframe of summary information of the dic samples
}
\description{
This provides a one-row dataframe (`data.table`) with information on the
 overall deviance, effective number of parameters (when `type` was pD) or
 optimism (when `type` was popt), and resulting penalised deviance.
}
\author{
Sam Clifford, \email{sj.clifford@gmail.com}
}
