% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tidier_stan.R
\name{mcmc_to_dt_stan}
\alias{mcmc_to_dt_stan}
\title{MCMC tidiers (draft) for STAN}
\usage{
mcmc_to_dt_stan(stan_model)
}
\arguments{
\item{stan_model}{a stan model}
}
\value{
a tidy dataframe
}
\description{
MCMC tidiers (draft) for STAN
}
\examples{
 mcmc_to_dt_stan(example_stan_model)


}
