% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_dt.R
\name{thin_dt}
\alias{thin_dt}
\title{thin_dt}
\usage{
thin_dt(dt, thin = 1)
}
\arguments{
\item{dt}{an object of class "data.table" returned from \code{mcmc_to_dt}}

\item{thin}{thinning interval}
}
\value{
a data.table dataframe
}
\description{
post-hoc thinning of MCMC chains which have been converted to a data.table
}
\examples{
library(coda)
data(line)
mcmc_dt <- mcmc_to_dt(line)
thin_dt(mcmc_dt, thin = 10)
thin_dt(mcmc_dt, thin = 2)
thin_dt(mcmc_dt, thin = 20)
}
