% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-mcmc-list.R
\name{tidy.mcmc.list}
\alias{tidy.mcmc.list}
\title{Return a tidy data summary of an MCMC object}
\usage{
\method{tidy}{mcmc.list}(x, conf_level = 0.95,
  aggregate_chains = FALSE, colnames = NULL, ...)
}
\arguments{
\item{x}{object of class "mcmc.list", as you would find with fitting a model
using \code{jags.model()}, and \code{coda.samples}.}

\item{conf_level}{level of the credible interval to be calculuated}

\item{aggregate_chains}{whether or not to summarise each parameter for each chain (FALSE, default) or to combine all chains (TRUE)}

\item{colnames}{which parameters we want from \code{mcmc_object}, if \code{NULL} then all
columns get selected}

\item{...}{extra arguments}
}
\value{
a data.table containing parameter summaries
}
\description{
\code{tidy.mcmc.list} is a function that behaves like those from \code{broom}. It takes
an mcmc.list object from \code{coda.samples} and return a data frame that
summarises each parameters with its mean and quantiles and returns the output
as a data.table object. This can be called as \code{tidy}. Currently summarises
over all chains.
}
\examples{
library(coda)
data(line)
tidy(line)
# Optionally ask for a subset of parameters with a vector of `colnames`,
# and summarise for each chain:
tidy(line,
     chain = TRUE,
     colnames=c("alpha"))
}
\author{
Sam Clifford, \email{sj.clifford@gmail.com}
}
